+++ config/custom-environment-variables.json
@@ -1,21 +1,21 @@
 {
   "passport": {
     "Google": {
-      "clientID": "GOOGLE_CLIENT_ID",
-      "clientSecret": "GOOGLE_CLIENT_SECRET",
-      "callbackURL": "GOOGLE_CALLBACK_URL"
+      "clientID": "",
+      "clientSecret": "",
+      "callbackURL": ""
     },
 
-    "Facebook": {
-      "clientID": "FACEBOOK_CLIENT_ID",
-      "clientSecret": "FACEBOOK_CLIENT_SECRET",
-      "callbackURL": "FACEBOOK_CALLBACK_URL"
+   "Facebook": {
+      "clientID": "",
+      "clientSecret": "",
+      "callbackURL": ""
     },
 
-    "Twitter": {
-      "consumerKey": "TWITTER_CLIENT_ID",
-      "consumerSecret": "TWITTER_CLIENT_SECRET",
-      "callbackURL": "TWITTER_CALLBACK_URL"
+   "Twitter": {
+      "consumerKey": "",
+      "consumerSecret": "",
+      "callbackURL": ""
     }
   }
 }
